Remove countries from indicator metadata dict
I don't think there's any point duplicating this info. The indicator IDs are
useful as they're not in the results and (I assume) they're unique to the WB,
but the ISO codes are already in the result keys, and you can easily lookup
countries by using get_countries() or pycountry.